refactor: reduce CharacterDetailView complexity and eliminate test duplication
Major architectural refactoring to address Codacy static analysis issues:

**Complexity Reduction:**
- Extract CharacterDetailView from 457 lines to 52 lines (88% reduction)
- Split monolithic component into 5 focused components:
  - CharacterOverview: header, core stats, classes, proficiency
  - CharacterStats: ability scores, saving throws, skills
  - CharacterEquipment: equipment display with badges
  - CharacterSpells: spells grouped by level with ordinal formatting
  - CharacterNotes: notes and backstory sections

**Code Organization:**
- Create character-utils.ts with shared D&D calculation functions
- Implement single responsibility principle for each component
- Maintain same external API and functionality

**Test Improvements:**
- Create shared-test-utils.tsx to eliminate duplication
- Reduce test code duplication from 18 clones to reusable utilities
- Fix test specificity issues with multiple text matches
- Implement common patterns: setupUserEvent, character builders, tab navigation

**Quality Assurance:**
- All tests passing (npm run test:ci ✓)
- Build successful (npm run build ✓)
- ESLint compliance (npm run lint ✓)
- Maintains 100% component test coverage

Addresses Codacy complexity warnings and code duplication issues.
